topK

+++ common/src/main/java/me/lzb/common/utils/ByteUtils.java
@@ -21,6 +21,6 @@ public class ByteUtils {
 
     public static int byteToInt(byte[] codes) {
         String s1 = byteToHexString(codes);
-        return Integer.valueOf(s1, 16).intValue();
+        return Integer.valueOf(s1, 16);
     }
 }
